+++ VSTS.psm1
@@ -11,7 +11,7 @@ $libs = Get-ChildItem `
     -Include '*.ps1' `
     -Recurse
 $libs.Foreach(
- {
+    {
         Write-Verbose -Message ('Importing the lib file {0}' -f $_.Fullname)
         . $_.Fullname
     }
@@ -41,7 +41,8 @@ $libs.Foreach(
 
     .PARAMETER Server
     The name of the VSTS or TFS Server to connect to.
-    For VSTS this will be 'visualstudio.com'. The default value
+    For VSTS this will be 'visualstudio.com', but for TFS
+    this should be set to the TFS server. The default value
     if this is not specified is 'visualstudio.com'.
 
     .PARAMETER HTTPS
@@ -53,29 +54,39 @@ $libs.Foreach(
 #>
 function New-VstsSession
 {
-    [CmdletBinding()]
+    [CmdletBinding(DefaultParameterSetName = 'VSTS')]
     [OutputType([PSCustomObject])]
     param
     (
-        [Parameter()]
+        [Parameter(Mandatory = $true, ParameterSetName = 'VSTS')]
+        [ValidateNotNullOrEmpty()]
         [String] $AccountName,
 
         [Parameter(Mandatory = $true)]
+        [ValidateNotNullOrEmpty()]
         [String] $User,
 
         [Parameter(Mandatory = $true)]
+        [ValidateNotNullOrEmpty()]
         [String] $Token,
 
         [Parameter()]
+        [ValidateNotNullOrEmpty()]
         [String] $Collection = 'DefaultCollection',
 
-        [Parameter()]
-        [String] $Server = 'visualstudio.com',
+        [Parameter(Mandatory = $true, ParameterSetName = 'TFS')]
+        [ValidateNotNullOrEmpty()]
+        [String] $Server,
 
         [Parameter()]
         [ValidateSet('HTTP', 'HTTPS')]
         [String] $Scheme = 'HTTPS'
     )
+
+    if ($PSCmdlet.ParameterSetName -eq 'VSTS')
+    {
+        $Server = 'visualstudio.com'
+    }
 
     [PSCustomObject] @{
         AccountName = $AccountName
@@ -154,6 +165,7 @@ function Get-VstsEndpointUri
         [Parameter(Mandatory = $true)]
         $Session,
 
+        [Parameter()]
         [String] $EndpointName
     )
 
@@ -352,8 +364,10 @@ function Get-VstsAuthorization
     [OutputType([String])]
     param
     (
+        [Parameter(Mandatory = $True)]
         [String] $User,
 
+        [Parameter(Mandatory = $True)]
         [String] $Token
     )
 
@@ -365,7 +379,7 @@ function Get-VstsAuthorization
     .SYNOPSIS
     Checks that a Guid is valid.
 
-    .PARAMETER Input
+    .PARAMETER Guid
     The Guid to validate.
 
     .OUTPUTS
@@ -378,9 +392,9 @@ function Test-Guid
     param
     (
         [Parameter(Mandatory = $True)]
-        $Input
+        [String] $Guid
     )
 
-    $Guid = [Guid]::Empty
-    [Guid]::TryParse($Input, [ref]$Guid)
+    $newGuid = [Guid]::Empty
+    [Guid]::TryParse($Guid, [ref]$newGuid)
 }